pyproject.toml
[tool.poetry]
name = "rest-api-django-advanced"
version = "08.02.2021"
description = "Code from REST API with Django Advanced"
authors = ["Paweł Benkowski <pawlo97.pb@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "2.1.3"
djangorestframework = "3.9.0"
flake8 = "3.6.0"
psycopg2 = "^2.8.6"
Pillow = "^8.1.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.2.2"
isort = "^5.7.0"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
